Finished splitting Body into sepreate Modal and Tweet components
