+++ tt-deploy/src/Success.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 
 const againbutton = <form action="http://google.com">
     <button type="link"> post another </button>
