+++ tt-deploy/src/Tweet.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 
 class Tweet extends React.Component{
     constructor(props){
